ch05-ex01-token-base-login/myblog/backend/pyproject.toml
[project]
name = "django-in-production-05-ex01"
version = "0.1.0"
description = "django-in-production 5章の発展型 token-base-login検証"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django==5.1.5",
    "psycopg2-binary==2.9.10",
    "djangorestframework==3.15.2",
    "django-rest-knox==5.0.2",
    "django-cors-headers==4.7.0",
]

[tool.uv]
link-mode = "copy"
dev-dependencies = [
    "ruff==0.9.2",
    "pytest==8.3.4",
    "pytest-django==4.9.0",
    "django-debug-toolbar==5.0.1",
]
